mcp-servers/calendar-mcp/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-secret-manager==2.20.0
